refactor(contracts): split the index.ts 万物筐 into 4 domain files (Stage 2)
packages/contracts/src/index.ts mixed 4 unrelated domains in one 400-line
file. Split into siblings, index.ts now a re-export barrel (flat public API
unchanged — all 69 consumers compile untouched):
- trade-decision.ts: live-order schemas + QUEUES/JOBS + admin/system enums
- public-api.ts: read-only public web-API shapes
- paper-trading.ts: paper-fill simulation logic
- rough-loop.ts: dev-tool task schemas (only services/rough-loop uses these)

Content moved verbatim (one type-identical tweak: PublicArtifactListItem.kind
uses Artifact["kind"] instead of the inline z.infer, avoiding a schema value
import in the types file). Behavior-preserving: contracts build + full-repo
typecheck green; 832/832 tests.
